+++ pwnlib/commandline/checksec.py
@@ -35,7 +35,10 @@ def main(args):
         return
 
     for f in files:
-        e = ELF(f.name)
+        try:
+            e = ELF(f.name)
+        except Exception as e:
+            print("{name}: {error}".format(name=f.name, error=e))
 
 if __name__ == '__main__':
     common.main(__file__)